bot.zoomlion.su/bot.zoomlion.su/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-chroma
langchain-huggingface
langchain
pydantic
unstructured[local-inference]
langchain-community
libmagic
python-magic
python-magic-bin
tesseract
poppler-utils